// src/app/services/models/voucher-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface VoucherDto {
  active?: boolean;
  arrivalDate?: string;
  description?: string;
  evictionDate?: string;
  hot?: boolean;
  hotelType?: string;
  id?: string;
  price?: number;
  title?: string;
  tourType?: string;
  transferType?: string;
}